Add subtyping rule [x has L:S..U => S <: U] and hmode to
control if we allow it (deeply).

Actually this is the same idea as in 0a9cf322 (exactly 11 months ago),
but this time I insisted more and it works for gDOT :-)
Note, though, that okhyp_to_nohyp currently relies on the fact that
subtyping never grows the env, which will not be the case any more
when we add existentials.